refactor: moved buttons to canvas folder
