+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewRadioQuestionButton/NewRadioQuestionButton.tsx
@@ -11,7 +11,7 @@ import CheckContainedIcon from '@core/shared/ui/icons/CheckContained'
 
 import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../../../__generated__/GetJourney'
 import { RadioQuestionBlockCreate } from '../../../../../../__generated__/RadioQuestionBlockCreate'
-import { Button } from '../../Button'
+import { Button } from '../../../ControlPanel/Button'
 
 export const RADIO_QUESTION_BLOCK_CREATE = gql`
   ${RADIO_QUESTION_FIELDS}
