+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewSignUpButton/NewSignUpButton.tsx
@@ -6,48 +6,40 @@ import type { TreeBlock } from '@core/journeys/ui/block'
 import { ActiveTab, useEditor } from '@core/journeys/ui/EditorProvider'
 import { ICON_FIELDS } from '@core/journeys/ui/Icon/iconFields'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
-import { TEXT_RESPONSE_FIELDS } from '@core/journeys/ui/TextResponse/textResponseFields'
-import TextInput1Icon from '@core/shared/ui/icons/TextInput1'
+import { SIGN_UP_FIELDS } from '@core/journeys/ui/SignUp/signUpFields'
+import Mail2Icon from '@core/shared/ui/icons/Mail2'
 
 import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../../../__generated__/GetJourney'
-import { TextResponseBlockCreate } from '../../../../../../__generated__/TextResponseBlockCreate'
-import { Button } from '../../Button'
+import { SignUpBlockCreate } from '../../../../../../__generated__/SignUpBlockCreate'
+import { Button } from '../../../ControlPanel/Button'
 
-export const TEXT_RESPONSE_BLOCK_CREATE = gql`
-  ${TEXT_RESPONSE_FIELDS}
+export const SIGN_UP_BLOCK_CREATE = gql`
+  ${SIGN_UP_FIELDS}
   ${ICON_FIELDS}
-  mutation TextResponseBlockCreate(
-    $input: TextResponseBlockCreateInput!
+  mutation SignUpBlockCreate(
+    $input: SignUpBlockCreateInput!
     $iconBlockCreateInput: IconBlockCreateInput!
     $id: ID!
     $journeyId: ID!
-    $updateInput: TextResponseBlockUpdateInput!
+    $updateInput: SignUpBlockUpdateInput!
   ) {
-    textResponseBlockCreate(input: $input) {
+    signUpBlockCreate(input: $input) {
       id
-      parentBlockId
-      parentOrder
-      ...TextResponseFields
     }
     submitIcon: iconBlockCreate(input: $iconBlockCreateInput) {
       id
       parentBlockId
       ...IconFields
     }
-    textResponseBlockUpdate(
-      id: $id
-      journeyId: $journeyId
-      input: $updateInput
-    ) {
-      ...TextResponseFields
+    signUpBlockUpdate(id: $id, journeyId: $journeyId, input: $updateInput) {
+      ...SignUpFields
     }
   }
 `
 
-export function NewTextResponseButton(): ReactElement {
-  const [textResponseBlockCreate] = useMutation<TextResponseBlockCreate>(
-    TEXT_RESPONSE_BLOCK_CREATE
-  )
+export function NewSignUpButton(): ReactElement {
+  const [signUpBlockCreate] =
+    useMutation<SignUpBlockCreate>(SIGN_UP_BLOCK_CREATE)
   const { journey } = useJourney()
   const {
     state: { selectedStep },
@@ -56,23 +48,21 @@ export function NewTextResponseButton(): ReactElement {
 
   const handleClick = async (): Promise<void> => {
     const id = uuidv4()
-    const submitIconId = uuidv4()
+    const submitId = uuidv4()
     const card = selectedStep?.children.find(
       (block) => block.__typename === 'CardBlock'
     ) as TreeBlock<CardBlock> | undefined
-
     if (card != null && journey != null) {
-      const { data } = await textResponseBlockCreate({
+      const { data } = await signUpBlockCreate({
         variables: {
           input: {
             id,
             journeyId: journey.id,
             parentBlockId: card.id,
-            label: 'Your answer here',
             submitLabel: 'Submit'
           },
           iconBlockCreateInput: {
-            id: submitIconId,
+            id: submitId,
             journeyId: journey.id,
             parentBlockId: id,
             name: null
@@ -80,11 +70,11 @@ export function NewTextResponseButton(): ReactElement {
           id,
           journeyId: journey.id,
           updateInput: {
-            submitIconId
+            submitIconId: submitId
           }
         },
         update(cache, { data }) {
-          if (data?.textResponseBlockUpdate != null) {
+          if (data?.signUpBlockUpdate != null) {
             cache.modify({
               id: cache.identify({ __typename: 'Journey', id: journey.id }),
               fields: {
@@ -97,16 +87,14 @@ export function NewTextResponseButton(): ReactElement {
                       }
                     `
                   })
-
                   const newBlockRef = cache.writeFragment({
-                    data: data.textResponseBlockUpdate,
+                    data: data.signUpBlockUpdate,
                     fragment: gql`
                       fragment NewBlock on Block {
                         id
                       }
                     `
                   })
-
                   return [
                     ...existingBlockRefs,
                     newBlockRef,
@@ -118,11 +106,10 @@ export function NewTextResponseButton(): ReactElement {
           }
         }
       })
-
-      if (data?.textResponseBlockCreate != null) {
+      if (data?.signUpBlockUpdate != null) {
         dispatch({
           type: 'SetSelectedBlockByIdAction',
-          id: data.textResponseBlockCreate.id
+          id: data.signUpBlockCreate.id
         })
         dispatch({
           type: 'SetActiveTabAction',
@@ -131,12 +118,13 @@ export function NewTextResponseButton(): ReactElement {
       }
     }
   }
+
   return (
     <Button
-      icon={<TextInput1Icon />}
-      value="Feedback"
+      icon={<Mail2Icon />}
+      value="Subscribe"
       onClick={handleClick}
-      testId="NewTextResponseButton"
+      testId="NewSignUpButton"
     />
   )
 }
